// frontend/src/api-generated/models/phieuSuaChuaBaoTriRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { ChiTietBaoTriLinhKienRequest } from './chiTietBaoTriLinhKienRequest';
import type { ChiTietBaoTriThietBiRequest } from './chiTietBaoTriThietBiRequest';

export interface PhieuSuaChuaBaoTriRequest {
  keHoachBaoTriId: number;
  thoiGianBatDau: string;
  thoiGianHoanThanhDuKien?: string;
  ghiChu?: string;
  danhSachThietBi?: ChiTietBaoTriThietBiRequest[];
  danhSachLinhKien?: ChiTietBaoTriLinhKienRequest[];
}